requirements.txt
fastapi
uvicorn
pandas
pytest
httpx
flake8
google-cloud-storage
autopep8